+++ osm_viewer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "osm_viewer"
+version = "0.1.0"
+authors = ["Dustin Carlino <dabreegster@gmail.com>"]
+edition = "2018"
+
+[features]
+default = ["map_gui/native", "widgetry/native-backend"]
+
+[dependencies]
+abstutil = { path = "../abstutil" }
+geom = { path = "../geom" }
+map_gui = { path = "../map_gui" }
+map_model = { path = "../map_model" }
+widgetry = { path = "../widgetry" }
